@@ -14,14 +14,12 @@
         }
     ],
     "require": {
-        "php": ">=7.1",
-        "symfony/framework-bundle": ">=2.7|>=3.3|4.*|5.*|6.*",
-        "sensio/framework-extra-bundle": "2.*|3.*|4.*|5.*|6.*",
+        "php": ">=8.1",
+        "symfony/framework-bundle": "6.*|7.*",
         "cocur/background-process": ">=0.7"
     },
     "require-dev": {
         "doctrine/orm": "^2.7",
-        "doctrine/annotations": "^1.5",
         "doctrine/cache": "^1.7",
         "doctrine/collections": "^1.5",
         "doctrine/instantiator": "^1.1",
