@@ -4,56 +4,40 @@ namespace Dtc\QueueBundle\Entity;
 
 use Doctrine\ORM\Mapping as ORM;
 
-/**
- * @ORM\Entity
- * @ORM\Table(name="dtc_queue_job_archive",indexes={
- *                  @ORM\Index(name="job_archive_status_idx", columns={"status"}),
- *                  @ORM\Index(name="job_archive_updated_at_idx", columns={"updated_at"})})
- */
+#[ORM\Entity]
+#[ORM\Table(name: 'dtc_queue_job_archive')]
+#[ORM\Index(columns: ['status'], name: 'job_archive_status_idx')]
+#[ORM\Index(columns: ['updated_at'], name: 'job_archive_updated_at_idx')]
 class JobArchive extends BaseJob
 {
-    /**
-     * @ORM\Column(name="id", type="bigint")
-     * @ORM\Id
-     */
+    #[ORM\Column(name: 'id', type: 'bigint')]
+    #[ORM\Id]
     protected $id;
 
     /**
      * When the job finished.
-     *
-     * @ORM\Column(name="finished_at", type="datetime", nullable=true)
      */
+    #[ORM\Column(name: 'finished_at', type: 'datetime', nullable: true)]
     protected $finishedAt;
 
-    /**
-     * @ORM\Column(name="elapsed", type="float", nullable=true)
-     */
+    #[ORM\Column(name: 'elapsed', type: 'float', nullable: true)]
     protected $elapsed;
 
     /**
      * When the job started.
-     *
-     * @ORM\Column(name="started_at", type="datetime", nullable=true)
      */
+    #[ORM\Column(name: 'started_at', type: 'datetime', nullable: true)]
     protected $startedAt;
 
-    /**
-     * @ORM\Column(name="when_us", type="decimal", precision=18, scale=0, nullable=true)
-     */
+    #[ORM\Column(name: 'when_us', type: 'decimal', precision: 18, scale: 0, nullable: true)]
     protected $whenUs;
 
-    /**
-     * @ORM\Column(name="priority", type="integer", nullable=true)
-     */
+    #[ORM\Column(name: 'priority', type: 'integer', nullable: true)]
     protected $priority;
 
-    /**
-     * @ORM\Column(name="expires_at", type="datetime", nullable=true)
-     */
+    #[ORM\Column(name: 'expires_at', type: 'datetime', nullable: true)]
     protected $expiresAt;
 
-    /**
-     * @ORM\Column(name="updated_at", type="datetime")
-     */
+    #[ORM\Column(name: 'updated_at', type: 'datetime')]
     protected $updatedAt;
 }
