@@ -5,27 +5,20 @@ namespace Dtc\QueueBundle\Entity;
 use Doctrine\ORM\Mapping as ORM;
 use Dtc\QueueBundle\Model\JobTiming as BaseJobTiming;
 
-/**
- * @ORM\Entity
- * @ORM\Table(name="dtc_queue_job_timing", indexes={@ORM\Index(name="job_timing_finished_at", columns={"status", "finished_at"})})
- */
+#[ORM\Entity]
+#[ORM\Table(name: 'dtc_queue_job_timing')]
+#[ORM\Index(columns: ['status', 'finished_at'], name: 'job_timing_finished_at')]
 class JobTiming extends BaseJobTiming
 {
-    /**
-     * @ORM\Column(name="id", type="bigint")
-     * @ORM\Id
-     * @ORM\GeneratedValue(strategy="AUTO")
-     */
+    #[ORM\Column(name: 'id', type: 'bigint')]
+    #[ORM\Id]
+    #[ORM\GeneratedValue(strategy: 'AUTO')]
     protected $id;
 
-    /**
-     * @ORM\Column(name="finished_at", type="datetime")
-     */
+    #[ORM\Column(name: 'finished_at', type: 'datetime')]
     protected $finishedAt;
 
-    /**
-     * @ORM\Column(name="status", type="integer")
-     */
+    #[ORM\Column(name: 'status', type: 'integer')]
     protected $status;
 
     /**
