@@ -5,14 +5,10 @@ namespace Dtc\QueueBundle\Document;
 use Doctrine\ODM\MongoDB\Mapping\Annotations as ODM;
 use Dtc\QueueBundle\Model\JobTiming as BaseJobTiming;
 
-/**
- * @ODM\Document(db="dtc_queue", collection="job_timing")
- */
+#[ODM\Document(db: 'dtc_queue', collection: 'job_timing')]
 class JobTiming extends BaseJobTiming
 {
-    /**
-     * @ODM\Id
-     */
+    #[ODM\Id]
     protected $id;
 
     /**
@@ -31,15 +27,11 @@ class JobTiming extends BaseJobTiming
         $this->id = $id;
     }
 
-    /**
-     * @ODM\Field(type="date")
-     * @ODM\Index(unique=false, order="asc")
-     */
+    #[ODM\Field(type: 'date')]
+    #[ODM\Index(unique: false, order: 'asc')]
     protected $finishedAt;
 
-    /**
-     * @ODM\Field(type="integer")
-     * @ODM\Index(unique=false, order="asc")
-     */
+    #[ODM\Field(type: 'integer')]
+    #[ODM\Index(unique: false, order: 'asc')]
     protected $status;
 }
