@@ -4,9 +4,7 @@ namespace Dtc\QueueBundle\Document;
 
 use Doctrine\ODM\MongoDB\Mapping\Annotations as ODM;
 
-/**
- * @ODM\Document(db="dtc_queue", collection="run")
- */
+#[ODM\Document(db: 'dtc_queue', collection: 'run')]
 class Run extends BaseRun
 {
 }
